+++ FakeNFT/Scenes /ShoppingCart/View/ViewControllers/PurchaseResultViewController.swift
@@ -50,7 +50,7 @@ class PurchaseResultViewController: UIViewController {
         setupResults()
         setupButton()
         setupCentre()
-        
+        setupLayout()
     }
     
     func setupResults() {
@@ -71,7 +71,12 @@ class PurchaseResultViewController: UIViewController {
     func setupCentre() {
         view.addSubview(centreImage)
         view.addSubview(centreLabel)
-        
+    }
+    
+    func setupButton() {
+        view.addSubview(resultButton)
+    }
+    func setupLayout() {
         NSLayoutConstraint.activate([
             centreImage.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 152),
             centreImage.centerXAnchor.constraint(equalTo: view.centerXAnchor),
@@ -85,22 +90,13 @@ class PurchaseResultViewController: UIViewController {
             
             centreLabel.centerXAnchor.constraint(equalTo: view.centerXAnchor),
             centreLabel.topAnchor.constraint(equalTo: centreImage.bottomAnchor, constant: 20),
-            centreLabel.heightAnchor.constraint(equalToConstant: 100)
-        ])
-        
-    }
-    
-    func setupButton() {
-        view.addSubview(resultButton)
-        
-        NSLayoutConstraint.activate([
+            centreLabel.heightAnchor.constraint(equalToConstant: 100),
             resultButton.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor, constant: -16),
             resultButton.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 16),
             resultButton.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16),
             resultButton.heightAnchor.constraint(equalToConstant: 60)
         ])
     }
-    
     @objc func paymentResultTapped() {
         
         if let tabBarController = presentingViewController?.presentingViewController as? UITabBarController {
